+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Warn when Azure PR metadata cannot provide a URL
+
 ## 3.7.0 - 2026-09-11
 
 ### Fixed